+++ Services/IDispatcherService.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace SwitchBlade.Services
+{
+    public interface IDispatcherService
+    {
+        void Invoke(Action action);
+    }
+}